+++ source/Setup/Database.php
@@ -27,6 +27,7 @@ use Exception;
 use PDO;
 use PDOException;
 use PDOStatement;
+use \OxidEsales\Facts\Facts;
 
 /**
  * Setup database manager class
@@ -307,9 +308,11 @@ class Database extends Core
 
         $oPdo = $this->getConnection();
 
-        $this->setIfDynamicPagesShouldBeUsed($oSession);
-
-        $blUseDynPages = isset($aParams["use_dyn_pages"]) ? $aParams["use_dyn_pages"] : $oSession->getSessionParam('use_dynamic_pages');
+        $blSendTechnicalInformationToOxid = true;
+        $facts = new Facts();
+        if ($facts->isCommunity()) {
+            $blSendTechnicalInformationToOxid = isset($aParams["send_technical_information_to_oxid"]) ? $aParams["send_technical_information_to_oxid"] : $oSession->getSessionParam('send_technical_information_to_oxid');
+        }
         $sLocationLang = isset($aParams["location_lang"]) ? $aParams["location_lang"] : $oSession->getSessionParam('location_lang');
         $blCheckForUpdates = isset($aParams["check_for_updates"]) ? $aParams["check_for_updates"] : $oSession->getSessionParam('check_for_updates');
         $sCountryLang = isset($aParams["country_lang"]) ? $aParams["country_lang"] : $oSession->getSessionParam('country_lang');
@@ -324,7 +327,7 @@ class Database extends Core
             $oPdo->exec("UPDATE oxuser SET oxcountryid = '$sCountryLang' where OXUSERNAME='admin'");
         }
 
-        $oPdo->exec("delete from oxconfig where oxvarname = 'blLoadDynContents'");
+        $oPdo->exec("delete from oxconfig where oxvarname = 'blSendTechnicalInformationToOxid'");
         $oPdo->exec("delete from oxconfig where oxvarname = 'sShopCountry'");
         $oPdo->exec("delete from oxconfig where oxvarname = 'blCheckForUpdates'");
         $oPdo->exec("delete from oxconfig where oxvarname = 'sDefaultLang'");
@@ -336,9 +339,9 @@ class Database extends Core
             [
                 'oxid' => $oUtils->generateUid(),
                 'shopId' => $sBaseShopId,
-                'name' => 'blLoadDynContents',
+                'name' => 'blSendTechnicalInformationToOxid',
                 'type' => 'bool',
-                'value' => $blUseDynPages
+                'value' => $blSendTechnicalInformationToOxid
             ]
         );
 
@@ -493,18 +496,5 @@ class Database extends Core
             "insert into oxconfig (oxid, oxshopid, oxvarname, oxvartype, oxvarvalue)
                              values('$sID', '$baseShopId', 'blSendShopDataToOxid', 'bool', ENCODE( '$blSendShopDataToOxid', '" . $configKey->sConfigKey . "'))"
         );
-    }
-
-    /**
-     * Set to session if dynamic pages should be used.
-     *
-     * @param Session $session
-     */
-    protected function setIfDynamicPagesShouldBeUsed($session)
-    {
-        // disabling usage of dynamic pages if shop country is international
-        if ($session->getSessionParam('location_lang') === null) {
-            $session->setSessionParam('use_dynamic_pages', 'false');
-        }
     }
 }