+++ source/Setup/tpl/_header.php
@@ -20,7 +20,8 @@
  * @version   OXID eShop CE
  */
 
-$editionSelector = new \OxidEsales\EshopCommunity\Core\Edition\EditionSelector();
+$facts = new \OxidEsales\Facts\Facts();
+
 ?>
 <!DOCTYPE html>
 <html>
@@ -40,19 +41,6 @@ $editionSelector = new \OxidEsales\EshopCommunity\Core\Edition\EditionSelector()
         }
     }
 
-<?php if (!$editionSelector->isEnterprise()) { ?>
-    function update_dynpages_checkbox()
-    {
-        sValue = document.forms[0].location_lang.value;
-        if ( sValue == '' ) {
-            document.getElementById('use_dynamic_pages_ckbox').style.display = 'none';
-            document.getElementById('use_dynamic_pages_desc').style.display = 'none';
-        } else {
-            document.getElementById('use_dynamic_pages_ckbox').style.display = '';
-            document.getElementById('use_dynamic_pages_desc').style.display = '';
-        }
-    }
-<?php } ?>
     /**
      * Replaces password type field into plain and vice versa
      */
@@ -82,15 +70,15 @@ $editionSelector = new \OxidEsales\EshopCommunity\Core\Edition\EditionSelector()
     <?php
         $iTabWidth = 147;
         $iSepWidth = 3;
-        if ($editionSelector->isEnterprise()) {
+        if ($facts->isEnterprise()) {
             $iTabCount = 7;
             $sHColor = '#006ab4';
         }
-        if ($editionSelector->isProfessional()) {
+        if ($facts->isProfessional()) {
             $iTabCount = 7;
             $sHColor = '#cd0210';
         }
-        if ($editionSelector->isCommunity()) {
+        if ($facts->isCommunity()) {
             $iTabCount = 6;
             $sHColor = '#ff3600';
         }