+++ tests/Acceptance/International/InternationalTest.php
@@ -437,7 +437,11 @@ class InternationalTest extends AcceptanceTestCase
             $this->frame("list");
             $this->openTab("Settings");
             $this->click("link=Administration");
-            $this->check("//input[@name='confbools[blLoadDynContents]' and @value='true']");
+
+            if ($this->getTestConfig()->getShopEdition() === 'CE') {
+                $this->check("//input[@name='confbools[blSendTechnicalInformationToOxid]' and @value='true']");
+            }
+
             $this->clickAndWaitFrame("save", "list");
             $this->frame("header");
             $this->assertElementPresent("link=Logout");