+++ CHANGELOG.md
@@ -26,6 +26,8 @@ facts component should be used: `\OxidEsales\Facts\Facts::getEdition()`.
 
 ### Removed
 
+- Dynamic pages related code including config option blLoadDynContents.
+
 ### Fixed
 - https://bugs.oxid-esales.com/view.php?id=6474 with PR#457
 - https://bugs.oxid-esales.com/view.php?id=6155 with PR#431